requirements.txt
opencv-python==4.9.0.80
numpy==1.24.4
pillow==10.3.0
scikit-image==0.21.0
streamlit
pillow